feat: Search user by submitting the form
Request the GitHub GraphQL API to search a user by query provided
from user on search form.
Missing pagination and some improvements on validation and
possible re-renderings.

FIXME: Found a weird behavior of API returning an empty node, when
query by "Camila", the first node seems to be an organization and
the response is an empty object `{}`. Requires investigation.